EH-747: Moved HankittavaYhteinenTutkinnonOsa.ts to a new folder
