+++ shared/models/YhteinenTutkinnonOsa/HankittavaYhteinenTutkinnonOsa.ts
@@ -1,8 +1,8 @@
 import { types } from "mobx-state-tree"
-import { YhteisenTutkinnonOsanOsaAlue } from "./YhteinenTutkinnonOsa/YhteisenTutkinnonOsanOsaAlue"
+import { YhteisenTutkinnonOsanOsaAlue } from "./YhteisenTutkinnonOsanOsaAlue"
 import { EPerusteetVastaus } from "models/EPerusteetVastaus"
-import { TutkinnonOsaType } from "./helpers/ShareTypes"
-import { EnrichTutkinnonOsaKoodiUri } from "./Enrichment/EnrichTutkinnonOsaKoodiUri"
+import { TutkinnonOsaType } from "../helpers/ShareTypes"
+import { EnrichTutkinnonOsaKoodiUri } from "../Enrichment/EnrichTutkinnonOsaKoodiUri"
 
 export const Model = types.model({
   moduleId: types.maybe(types.string),
